v1 minus sendgrid, s3, lead mailer

+++ db/migrate/20160110042103_create_workout_times.rb
@@ -2,7 +2,7 @@ class CreateWorkoutTimes < ActiveRecord::Migration
   def change
     create_table :workout_times do |t|
       t.date :date, default: Date.today.to_s, null: false
-      t.integer :time, default: 0, null: false
+      t.float :time, default: 0, null: false
       t.belongs_to :user
 
       t.timestamps null: false
